+++ build.sh
@@ -2,12 +2,13 @@
 
 rm -rf lib/
 
-cp ./node_modules/ts-transformer-interface/transformer.ts ./transform-interface.ts
-cp ./node_modules/ts-transformer-interface/runtime-schema.ts ./runtime-schema.ts
-cp ./node_modules/ts-transformer-decoder-cast/transformer.ts ./transform-request.ts
+cp ./node_modules/ts-transformer-interface/transformer.d.ts ./transform-interface.d.ts
+cp ./node_modules/ts-transformer-interface/transformer.js ./transform-interface.js
 
-cat runtime-schema.ts | sed 's/index.d/index/' > replaced
-mv replaced runtime-schema.ts
+cp ./node_modules/ts-transformer-interface/runtime-schema.d.ts ./runtime-schema.d.ts
+cp ./node_modules/ts-transformer-interface/runtime-schema.js ./runtime-schema.js
+
+cp ./node_modules/ts-transformer-decoder-cast/transformer.ts ./transform-request.ts
 
 tsc
 npx pretty-quick